v0.3.3 - Section Navigation, Scroll & Heading
Added a section navigation. The user can just click the "prev" or "next" button (with arrows) and the section will fade-in.

The section element is now scrollable, so when the content is too long it still can be accessed,

Added a section heading and subheading design.

Added the input and "add element" button design, may change it later if needed.