v0.3.4 - Mobile Nav & Section Heading Fixes
Fixed a transition when using the section nav (arrows) to change the section. Previously, it would fade-in and out the icons resulting in some visual mess and icon "jumping".

Fixed the section heading gradient. It was stretched to fit the 100% width of the section. The actual text had only the middle part of the gradient.
After the fix the gradient looks correct.