v0.4.2 - Input Validation, PDF Print
Added input validation.
Some of the inputs can't be empty. If they are, the PDF print button won't have any action. The input and the nav section element will change color to indicate the empty inputs.

Added a PDF print function to the button, so the PDF can be downloaded or printed.