v0.5.0 - SCSS, Profile Section Styles, Nav Fix
Added SASS (SCSS).

Changed the Profile section a bit and added media queries.
There is always 1 link to add, so the user can click "+" button if they want more links (and not if they want only 1 link like before).

Fixed some small bugs with nav.
The icon was not sized properly on really small screen sizes.
There was a gap between the text and the icon (a place where nothing happened when clicked).