v0.6.0 - QR Code, Links Improvement, Input Fix
Added a QR Code subsection.
The user can drag & drop a file or upload it with a button (the same as photo upload) and give the QR code a name.
To achieve that, the file upload script was modified to accept parameters and call the script from a single function, so it's possible to easily create multiple drag & drop elements.

Improved the Links subsection.
Now, the first link also has the "delete" button, but instead of removing the input, it only removes the input value (from the PDF section too).
The link elements are appended to a different div, so the code is more readable and easy to change.

Fixed a small bug where the inputs were white in dark mode.