v0.7.0 - Creating Elements, Link Elements, Placeholders
Changed the script for creating elements (e.g. Link Elements).
Now, the element can be created by calling a function and using an object as a parameter. The object contains the info needed to create various elements. There are functions for a div, label, input and text area.

It's more readable and the code redistribution helps to arrange everything in a function (no long chains of adding classes, IDs, data attributes, etc.).

Changed the link elements code to be created with the new script and to be able to remove the first element.

Added placeholders to the input elements for better UX.