v0.8.0 - Work Experience
Added the Work Experience section.

The user can add a few job positions.
When the user clicks the add button, the section automatically expands. Only one section at a time can be expanded (to not make a visual clutter, better UX).
The user can expand or collapse the section by clicking on the section head.

The section head includes a job title (sourced from the input value, helps to identify the job after collapsing the section), delete section button and expand/collapse section indicator.

There are a few inputs, including start/end date, current position checkbox and a job description.

The user has the ability to make the text bold/italic/underline in the job description and there is an option to generate a list (numbers or bullet points).

Every section can be removed.

Made one of the box-shadow types more visible in dark mode.