v0.9.0 - Education Section, List Collapse Fix
Added an Education section.
The code is similar to the experience section, but there are some minor changes.

The section inputs are changed to match the education related info.
The list title is made of the degree and school name.

Additionally, added placeholders to the education and experience section.

Fixed a small bug where the list wouldn't collapse when clicking on a specific list head part.