v0.15.1 - PDF Color Settings
Added some inputs to change the PDF colors, like aside section or the main color.

The user can reset the color to the original value.

The user can also change the aside text color to achieve a sufficient contrast with the background.
If the contrast is too low, there will be a warning message displayed.