v0.19.0 - Design Section, Many Other Changes
Minor changes to the label elements across the whole program (smaller font + more faded color).

Rearranged some subsections in the "Design" section.

Made layout changes to the "Design" section elements and applied styles to the inputs to match the general program design.

Added media queries to the "Design" section.

Added a switch for the photo border. The user can choose if the border should be the main or the text color.

Changed the text to background contrast warning message. Now, there is a message "Choose your color palette." that changes to the warning message if necessary.

Changed the switch input info behavior. The text is not changing, instead there are 2 side messages to better match the input functionality.

The select input for the typefaces now has the appropriate typefaces that the option sets.

Further implementation of the design input user changes.

Fixed a bug where some entered data wouldn't be displayed in the PDF view.

Fixed label warnings - "No label associated with a form field".

Other small bug fixes and design changes.