v0.20.0 - PDF Expanded View
Fixed the PDF expanded view, so it's scrollable, keeps the aspect ratio.
The size is adjusted for every screen resolution, so it looks good on mobile as well.

On desktop and horizontal mobile (on a side), the user can scroll vertically to take advantage of the wide screen.
On mobile (vertical screen), the user is able to scroll horizontally to take advantage of device's high, vertical screen.

Additionally, changed the placement of the close PDF expanded view button on mobile, so it's easier to reach (moved it from top/right to bottom/right).