v0.21.0 - PDF Page Dimensions, Aside Width
Added a select dropdown that allows the user to select different page sizes or aspect ratio (could be useful for online presentations, social media, etc.).

Added a range input for aside width. It can make the resume look better if there are a lot or only a few aside elements.