v0.22.0 - PDF View Changes, More User Inputs
Made multiple PDF design changes for better readability, visual experience and consistency, such as paddings, gaps, text appearance, etc.
Fixed multiple PDF view layout issues.

Added a switch for the aside placement - left or right side of the page.
Added a range slider input for changing the page padding.

Added a Text Align select element that allows the user to choose the alignment of the long-form texts (summart, description, etc.).
Added a Text Balance switch - it disables/enables the new CSS text-wrap:balance; feature for the long form texts (enabled by default).

Made the Professional Summary section into a separate section in the PDF view. Also moved it up in the Profile section to better match the PDF view layout order.
Added the text edit options to the Professional Summary section.
Fixed an issue where if the (for example) bold button was clicked (became active), then the user selected normal text and clicked the button once more (button not active), the normal text would become bold.
Made it so when the user is pasting something into the description textarea, only the plain text is inserted. Before that, the formatting would be pasted too, resulting in different typefaces, etc. that were visible in the PDF view.

When the user adds a list point (number or bullet) in the description box, like in the experience section, etc, the focus is automatically set to be after the last character (so after the added list point).
Improved the section design elements in the PDF view (dots on the line on the left side) - there were small inconsistencies with the dot alignment.

Fixed some small issues with the PDF view sections having influence on the other sections if they are empty.
Also, removed the section design element when the section is added, but is still empty.

Changed and simplified some code to be more readable and not repetetive.
Implemented multiple other design and layout changes.