v0.22.1 - Settings, Optimization, Design Fixes
Greatly improved optimization (nav and sections) by changing transitions to use hardware acceleration. When there was a lot of content added the animations would be choppy, now they look perfectly smooth.

Updated the "Settings & Info" section (design and content).

Changed the space between text to be corelated with the text / heading font size.

Centered the switch inputs on mobile for better UX.

Fixed some issues with the address section when switching to text instead of icons.