v1.0.6 - Lang Lvl, QR Code Fixes
Fixed a problem with some of the language level designs, where it wouldn't look good (CSS rendering problem).
Additionally, changed the link title color to be less intense to differentiate it from the URL.

Fixed a small bug where there was a gap between aside Links section and the QR Code section, even if there was no QR Code title or file.
Now, a title or an uploaded file is required to affect other sections.