v1.1.0 - Additional Info Section, Text-Wrap Bug
Added an Additional Info Section (Custom Section) in the References & Info.
The user can change the section heading and input whatever content they want - functioning similar to the summary and description text area (there are buttons for the list, bold text, etc.).

Additionally, the user can choose to snap the section to the bottom of the page and make the text smaller than any other text in the document (especially useful for privacy data consent and other long paragraphs). The settings can be found in the Design section.

Fixed a small bug where the pasted text would not have the text-wrap property set to balance (pasted text is in a <span> tag, but the text-wrap was on the parent tag).

Adjusted the background color in the PDF view, so when the program is in the dark-mode the corners don't have a slightly light part (like a shadow).

Added a folder (gitignore) for the design prototypes of the program.