v1.2.0 - Clickable Links
The links added by the user in the Links section are now clickable in the PDF document.
The "https://" part is added automatically if omitted, so the link works correctly no matter how it looks like.

There is a popup info in the Links section to inform the user that if they omit the "https://" part of the link it will be added automatically and if the link has "http://" they should add it or it might not work correctly.