v1.3.1 - Nav Animation Improvements & Fixes
Improved the animation for the nav icons (Profile, Experience).
Changed the experience icon to display documents when opened.

Fixed a small issue with the courses icon where the checkmark was not behaving as expected.

Made the "shake" animation as default for every icon, but each one has a unique feature, so every animation is different.