v1.4.0 - Image, Layout, Headings, Dates
Updated the resume photo to use various aspect ratios - from 3/4 to 4/3.
Previously, it was fixed at 1/1 only.
The photo is using all of the available space.
If it does not fit, the min/max aspect ratio is used, and the photo is cropped to cover the available space.

Adjusted the photo corner rounding for the different aspect ratios.
Now, it's calculated based on the photo width, instead of percentage that only works for square elements.

Added toggles for the skills and hobbies sections layout.
Now, the user can switch between wrapping the tiles (row-wrap) and placing them in a single column.

Implemented a maximum width for the subsection headings (job title, etc.) so the date always fits.
Added "text-wrap: pretty" for the subsection headings so that if the text moves to the next line, it is more evenly distributed.

Added the "contenteditable" property to the dates so they can be changed directly in the PDF view.

Implemented minor design changes.